[Spring 지하철 노선도 - 3단계] 알파(배철환) 미션 제출합니다. (#331)
* docs : 기능요구사항 추가

* docs : 도메인 추가 및 기능요구사항 추가

* feat : 구간을 추가하고 삭제하는 기능을 담당하는 Stations 추가

* feat : 한 노선의 상행 종점과 하행 종점을 찾는 기능 추가

* feat : SectionDao의 기능 구현

* feat : 구간을 추가하고 삭제하는 SectionService의 기능 추가

* feat : LineService에 노선 추가, 삭제, 수정, 조회, 목록 기능 추가

* feat : StationService에 역 추가, 삭제, 조회 기능 추가

* feat : SectionController의 기능 추가

* feat : LineController의 기능 추가

* refactor : 구간 삭제와 구간 삽입의 비즈니스 로직을 담당하는 Sections 추가

* refactor : 삭제된 MetroManager를 의존하던 클래스와 테스트 케이스 수정

* refactor : 역을 삭제하는 경우 내부 로직 수정

* refactor : 클래스명 변경

* refactor : SectionDao의 메소드명 변경

* refactor : DB에서 특정 데이터가 있는지 조회할 때 count가 아닌 exist로 수정

* refactor : 테스트 코드 내의 사용하지 않는 지역변수 제거

* refactor : 커스텀 예외가 아닌 예외에 대한 Response를 추가

+++ src/main/java/wooteco/subway/dto/SectionRequest.java
@@ -3,12 +3,12 @@ package wooteco.subway.dto;
 public class SectionRequest {
     private Long upStationId;
     private Long downStationId;
-    private int distance;
+    private Long distance;
 
     public SectionRequest() {
     }
 
-    public SectionRequest(Long upStationId, Long downStationId, int distance) {
+    public SectionRequest(Long upStationId, Long downStationId, Long distance) {
         this.upStationId = upStationId;
         this.downStationId = downStationId;
         this.distance = distance;
@@ -22,7 +22,7 @@ public class SectionRequest {
         return downStationId;
     }
 
-    public int getDistance() {
+    public Long getDistance() {
         return distance;
     }
 }
